+++ tests/utils_tests.cpp
@@ -67,15 +67,15 @@ TEST_CASE("test check shape dimension") {
   CHECK_EQ(check_shape_dim(-4), -4);
   CHECK_EQ(check_shape_dim(0), 0);
   CHECK_EQ(check_shape_dim(12), 12);
-  CHECK_EQ(check_shape_dim(static_cast<ssize_t>(dim_min)), dim_min);
-  CHECK_EQ(check_shape_dim(static_cast<ssize_t>(dim_max)), dim_max);
+  CHECK_EQ(check_shape_dim(static_cast<int64_t>(dim_min)), dim_min);
+  CHECK_EQ(check_shape_dim(static_cast<int64_t>(dim_max)), dim_max);
   CHECK_EQ(check_shape_dim(static_cast<size_t>(0)), 0);
   CHECK_EQ(check_shape_dim(static_cast<size_t>(dim_max)), dim_max);
   CHECK_THROWS_AS(
-      check_shape_dim(static_cast<ssize_t>(dim_min) - 1),
+      check_shape_dim(static_cast<int64_t>(dim_min) - 1),
       std::invalid_argument);
   CHECK_THROWS_AS(
-      check_shape_dim(static_cast<ssize_t>(dim_max) + 1),
+      check_shape_dim(static_cast<int64_t>(dim_max) + 1),
       std::invalid_argument);
   CHECK_THROWS_AS(
       check_shape_dim(static_cast<size_t>(dim_max) + 1), std::invalid_argument);
